fish: set $Getopt::Std::STANDARD_HELP_VERSION to true
According to the Getopt::Std manpage, it is strongly recommended to
set this variable.

Without this option, calling "fish --version" does not do the right
thing --- a user expects either that the version is printed, or that
the option is unknown and the call fails.

See also https://github.com/plicease/Shell-Config-Generate/issues/10
where this showed up as a problem.