Fix issue with NULL filehandle in patch.
Currently patch saves an reference to a localized NULL filehandle in the
object hash in its bless() method. But when the bless() method returns
the original value of *NULL is restored (which is empty). Meaning that
when patch later calls apply() and tries to print to the  NULL handle,
we get a warning about trying to print to a closed filehandle.

The correct behavior should be to store a copy of *NULL (not a
reference *NULL)  or to not localize *NULL in the first place..